--- msg_package_template/msg/Template.msg
@@ -1,6 +1,0 @@
-uint8 UNKNOWN = 0
-uint8 PLANE = 1
-uint8 SPHERE = 2
-
-uint8 type
-msg_package_template/ModelCoefficients coefficients
