--- msg_package_template/package.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>msg_package_template</name>
-  <version>0.0.0</version>
-  <description>The msg_package_template package</description>
-
-  <maintainer email="ashwins@iitk.ac.in">Ashwin Shenai</maintainer>
-  <maintainer email="gajendarnagar02@gmail.com">Gajendra Nagar</maintainer>
-
-  <license>TODO</license>
-
-  <!-- The *depend tags are used to specify dependencies -->
-  <!-- Dependencies can be catkin packages or system dependencies -->
-  <!-- Examples: -->
-  <!-- Use depend as a shortcut for packages that are both build and exec dependencies -->
-  <!--   <depend>roscpp</depend> -->
-  <!--   Note that this is equivalent to the following: -->
-  <!--   <build_depend>roscpp</build_depend> -->
-  <!--   <exec_depend>roscpp</exec_depend> -->
-  <!-- Use build_depend for packages you need at compile time: -->
-  <!--   <build_depend>message_generation</build_depend> -->
-  <!-- Use build_export_depend for packages you need in order to build against this package: -->
-  <!--   <build_export_depend>message_generation</build_export_depend> -->
-  <!-- Use buildtool_depend for build tool packages: -->
-  <!--   <buildtool_depend>catkin</buildtool_depend> -->
-  <!-- Use exec_depend for packages you need at runtime: -->
-  <!--   <exec_depend>message_runtime</exec_depend> -->
-  <!-- Use test_depend for packages you need only for testing: -->
-  <!--   <test_depend>gtest</test_depend> -->
-  <!-- Use doc_depend for packages you need only for building documentation: -->
-  <!--   <doc_depend>doxygen</doc_depend> -->
-
-  <buildtool_depend>catkin</buildtool_depend>
-  <buildtool_depend>catkin_simple</buildtool_depend>
-  <buildtool_depend>message_generation</buildtool_depend>
-
-  <depend>message_generation</depend>
-  <depend>std_msgs</depend>
-
-</package>
