+++ pcl_ransac_msgs/msg/CloudModel.msg
@@ -1,0 +1,6 @@
+uint8 UNKNOWN = 0
+uint8 PLANE = 1
+uint8 SPHERE = 2
+
+uint8 type
+pcl_ransac_msgs/ModelCoefficients coefficients
