+++ pcl_ransac_ros/launch/pcl_ransac.launch
@@ -1,0 +1,10 @@
+<?xml version="1.0"?>
+<launch>
+
+    <node pkg="pcl_ransac_ros" type="pcl_ransac_ros_node" name="pcl_ransac_ros_node" output="screen">
+        <rosparam command="load" file="$(find pcl_ransac_ros)/cfg/PCGenParams.yaml" />
+    </node>
+
+    <node pkg="rviz" type="rviz" name="rviz" args="-d $(find pcl_ransac_ros)/rviz/visualizer.rviz" />
+
+</launch>
