--- ros_package_template/launch/template.launch
@@ -1,10 +1,0 @@
-<?xml version="1.0"?>
-<launch>
-
-    <node pkg="ros_package_template" type="ros_package_template_node" name="ros_package_template_node" output="screen">
-        <rosparam command="load" file="$(find ros_package_template)/cfg/template.yaml" />
-    </node>
-
-    <node pkg="rviz" type="rviz" name="rviz" args="-d $(find ros_package_template)/rviz/template.rviz" />
-
-</launch>
